Rename time module to time_ to avoid native module
