+++ external-import/zerofox/src/time_/connectorInterval.py
